requirements.txt
openai-whisper
static-ffmpeg
numpy 